Moved element methods to new class + new type
Moved element methods away from mapdl_functions to a new class
_MapdlElementCommands. Some methods may have been missed. Amended
`self.run` to remove possibility of returning int and factored it into
the element_commands methods. Created a dedicated mapdl_types file and
added the MapdlInt type, which can be str or int.
