+++ ansys/mapdl/core/mapdl_types.py
@@ -1,0 +1,3 @@
+from typing import Optional, Union
+
+MapdlInt = Union[str, int]  # mapdl integers can either be strings or integers. "1" and 1 are equally valid
